Feat: 상세페이지 통합
