--- src/component/Detail/css/index.css
@@ -1,3 +1,0 @@
-.detailContainer {
-  background-color: white;
-}
