--- src/component/Detail/index.js
@@ -1,18 +1,0 @@
-import React from 'react'
-import './css/index.css'
-
-import Social from './Social'
-import Contents from './Contents'
-import Comment from './Comment'
-
-function Index() {
-  return (
-    <div className="detailContainer">
-      <Social />
-      <Contents />
-      <Comment />
-    </div>
-  )
-}
-
-export default Index
