+++ src/components/Detail/css/Contents.css
