+++ src/components/Detail/css/Social.css
