+++ src/components/Detail/index.js
@@ -3,6 +3,9 @@ import './css/index.css'
 import Header from './Header'
 import ProductInfo from './ProductInfo'
 import Image from './Image'
+import Social from './Social'
+import Contents from './Contents'
+import Comment from './Comment'
 
 export default function Index() {
   return (
@@ -10,6 +13,9 @@ export default function Index() {
       <Header />
       <ProductInfo />
       <Image />
+      <Social />
+      <Contents />
+      <Comment />
     </div>
   )
 }
